Storing configuration in variables
also, variables changes are directly impacted on the map.
