+++ src/Components/Configs/ExitConfig.ts
@@ -2,7 +2,7 @@
  * Declare and export the exitConfig interface
  */
 
-export interface exitConfig {
+export interface ExitConfig {
     nameSpace : string,
     active: boolean,
     url : string,
