+++ src/Stores/OpenCoWebsiteStore.ts
@@ -1,11 +1,11 @@
 import { writable } from "svelte/store";
-import type { openCoWebSiteConfig } from "../Components/Configs/OpenCoWebSiteConfig";
+import type { OpenCoWebSiteConfig } from "../Components/Configs/OpenCoWebSiteConfig";
 
 /**
  * A store that contains an array of the configuration of the OpenCoWebSite of the map
  */
 
-export const OpenCoWebsiteStore = writable<openCoWebSiteConfig>(
+export const OpenCoWebsiteStore = writable<OpenCoWebSiteConfig>(
     {
         nameSpace: "reception",
         url: '',
