+++ src/iframe.ts
@@ -1,7 +1,13 @@
-/// <reference path="../../../workadventure-saas/workadventure/front/dist/src/iframe_api.d.ts" />
-// @ts-ignore
+/// <reference path="../../../workadventure.src/workadventure/front/dist/src/iframe_api.d.ts" />
+
 import App from "./Components/App.svelte";
 import "../style/style.scss";
+import {mapVariableToStore} from "./Stores/VariableMapper";
+import {MeetingRoomStore} from "./Stores/MeetingRoomStore";
+
+WA.onInit().then(() => {
+    //mapVariableToStore('meetingRooms', MeetingRoomStore);
+});
 
 const app = new App({
     target: document.body,
